deleted atLeastOneNonNullParameter method and changed entity id from int to long

+++ stage #3/java/module #2. REST API Basics/task02/repository/src/main/java/by/epam/esm/dao/CertificateSqlQueryBuilderImpl.java
@@ -91,7 +91,7 @@ public class CertificateSqlQueryBuilderImpl implements CertificateSqlQueryBuilde
         query.deleteCharAt(query.lastIndexOf(COMMA));
     }
 
-    private void addWhereIdEqual(Integer id) {
+    private void addWhereIdEqual(Long id) {
         query.append(WHERE + ID + EQUALLY + EXCEPTED_VALUE);
         params.add(id);
     }
