+++ stage #3/java/module #2. REST API Basics/task02/service/src/main/java/by/epam/esm/service/GiftCertificateService.java
@@ -12,7 +12,7 @@ import java.util.List;
  * @author Aliaksei Tkachuk
  * @version 1.0
  */
-public interface GiftCertificateService extends BaseService<GiftCertificateDto,Integer> {
+public interface GiftCertificateService extends BaseService<GiftCertificateDto,Long> {
     /**
      * method findAll
      * method for find all certificates dto
